+++ fossic-node/src/lib.rs
@@ -1,14 +1,16 @@
+mod iters;
 mod store;
 mod subscriptions;
 mod types;
 
 use napi_derive::napi;
 
+pub use iters::{FossicCausationIter, FossicCorrelationIter, FossicRangeIter};
 pub use store::Store;
 pub use subscriptions::FossicSubscription;
 pub use types::{
-    AppendJs, BranchInfoJs, CreateBranchJs, EventId, OpenOptionsJs, ReadQueryJs, StreamInfoJs,
-    StoredEventJs, SubscribeQueryJs,
+    AppendJs, BranchInfoJs, CreateBranchJs, EventId, OpenOptionsJs, ReadOutcomeJs, ReadQueryJs,
+    SamplingModeJs, StreamInfoJs, StoredEventJs, SubscribeQueryJs, TruncationCursorJs,
 };
 
 /// Fossic v1 napi-rs Node.js binding.